requirements.txt
apache-airflow==2.8.1
black==24.4.2
pylint==3.1.1
isort==5.13.2
pandas==2.2.2